Removed all files containing the illegal '?' character
During scraping any files that the old websites used were scraped and saved as is
However, in NTFS the '?' character is illegal, but during scraping these were saved meaning that windows users cannot clone this repository to make changes.
Looking through the files it seems that pages that contain a '?' also have a sibling where the '?' is replaced with a '@' (which is not illegal) and it also seems the pages all link to these siblings and not the originals.
As such it should be safe to remove these files with the '?', if not this commit can safely be removed.